% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{cormid}
\alias{cormid}
\title{Mid-correction of cummulative weights of a sample}
\usage{
cormid(xs)
}
\arguments{
\item{xs}{object from class \code{wtsamp} containing the sample
and weights for components' distributions.}
}
\value{
object from class \code{wtsamp} whose cummulative weights are
mid-corrected and individual weights are set to NULL.
}
\description{
\code{cormid} calculates cummulative weights as the mean
of sup- and inf-corrected weights
  (mid-corrected weights).
}
\details{
If cummulative weights are NULL they will be calculated
from the individual ones by \code{indiv2cum}.
}
\examples{
p <- genunifp(15,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
F1 <- edfgen(xs,1) # minimax the estimate for 1-st component
xs_mid <- cormid(xs)
F1_mid <- edfgen(xs_mid,1) # mid-corrected estimate for 1-st component
F1(0)
F1_mid(0)
plot(F1,-3,3)
curve(F1_mid,col="red",,lty="dashed",add=TRUE)
}
\seealso{
Maiboroda R.  and  Kubaichuk O.
Asymptotic normality of improved weighted empirical distribution functions.
Theor. Probability and Math. Statist. 69 (2004), 95-102
}
