% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{corsup}
\alias{corsup}
\title{Sup-correction of cummulative weights of a sample}
\usage{
corsup(xs)
}
\arguments{
\item{xs}{object from class \code{wtsamp} containing the sample
and weights for components' distributions.}
}
\value{
object from class \code{wtsamp} whose cummulative weights are
sup-corrected and individual weights are set to NULL.
}
\description{
\code{corsup} calculates cummulative weights for
the upper nondecreasing envelope of the
CDFs of mixture components (sup-corrected weights). The weights are
truncated by 1.
}
\details{
If cummulative weights are NULL they will be calculated
from the individual ones by \code{indiv2cum}.
}
\examples{
p <- genunifp(15,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
F1 <- edfgen(xs,1) # minimax the estimate for 1-st component
xs_sup <- corsup(xs)
F1_sup <- edfgen(xs_sup,1) # sup-corrected estimate for 1-st component
F1(0)
F1_sup(0)
plot(F1,-3,3)
curve(F1_sup,col="red",,lty="dashed",add=TRUE)
}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.
}
