% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{cumm2indiv}
\alias{cumm2indiv}
\title{Calculates individual weights.}
\usage{
cumm2indiv(xs)
}
\arguments{
\item{xs}{a \code{wtsamp} or \code{wtcens} object representing
a sample with distributions of different components.}
}
\value{
an object of the same class as \code{xs} with recalculated
individual weights.
}
\description{
\code{cumm2indiv} calculates differences with lag 1 from
cummulative weights and put them into individual weights
attribute.
}
\examples{
set.seed(3)
p <- genunifp(10,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
ys <- indiv2cumm(xs) # create cummulative weights
xs1 <- cumm2indiv(ys) #xs1 is the same as xs
}
