% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{densgen}
\alias{densgen}
\title{Generator of kernel density estimator for mixture components.}
\usage{
densgen(xs, m, Kern = Epanechn)
}
\arguments{
\item{xs}{a \code{wtsamp}  object representing
a sample with distributions of different components.}

\item{m}{a number of component whose density is estimated.}

\item{Kern}{a function which calculates the kernel values
at given points (must be vectorized).}
}
\value{
a function
\code{f(x,h)} which calculates the estimate at points given in the
vector \code{x} with the bandwidth  \code{h}.
}
\description{
\code{densgen} generates a function which calculates
a kernel density estimate for a prescribed mixture component
at a given set of points.
}
\examples{
set.seed(3)
p <- genunifp(1000,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
f<-densgen(xs,1) # create the estimator
f(c(0,1),1)      # calculate the estimate at two points
curve(f(x,1),-3,3) # plot the graph (estimates N(0,1) density)

}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.
}
