% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{edfgen}
\alias{edfgen}
\title{Generator of weighted empirical CDS.}
\usage{
edfgen(xs, m)
}
\arguments{
\item{xs}{a \code{wtsamp} or \code{wtcens} object representing
a sample with distributions of different components.}

\item{m}{number of the component whose CDF is estimated.}
}
\value{
a function with the call \code{f(t)} where \code{t} is
the vector of points at which the estimate is calculated.  \code{f(t)}
returns the vector of estimates.
}
\description{
Generator of weighted empirical CDS.
}
\examples{
set.seed(3)
p <- genunifp(1000,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
F1<-edfgen(xs,1) # generate the estimate for 1-st component
F1(0)            # 0.5289866 approximately 0.5
plot(F1,-3,3) # plot the estimate (approx. standard normal CDF )
}
