% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescript.R
\docType{data}
\name{EIT2016}
\alias{EIT2016}
\title{Results of Ukrainian External Independent Testing 2016}
\format{A data frame with 94680 and 3 variables. Each row corresponds
to one person's test results. The variables are:
\describe{
 \item{obl}{code of region of Ukraine (oblast) where the examenee
   finished the high school (the codes are the same as in ukrelect2014
   data file)}
  \item{ukr}{result of test on Ukrainian language and literature}
 \item{math}{result of test on mathematics}
}}
\source{
Official site of Ukrainian Center for
Educational Quality Assessment
\url{https://zno.testportal.com.ua/stat/2016}
}
\usage{
EIT2016
}
\description{
A dataset containing results of tests in Mathematics and
Ukrainian language and literature at External Independent Testing
2016 in Ukraine. The data on persons which passed both Math and Ukrainian
tests are included only.
}
\keyword{datasets}
