% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{gencensg}
\alias{gencensg}
\title{Censored sample from
mixture of gamma distributions with varying concentrations.}
\usage{
gencensg(p, shape, rate, shapec, ratec)
}
\arguments{
\item{p}{matrix (or data frame) of mixing probabilities.
with rows corresponding to subjects.
and columns coresponding to the mixture components.}

\item{shape}{vector of shape parameters for gamma distributions of
non-censored components.}

\item{rate}{vector of rate parameters for gamma distributions of
non-censored components.}

\item{shapec}{vector of shape parameters for gamma distributions of
components' censors.}

\item{ratec}{vector of rate parameters for gamma distributions of
components' censors.}
}
\value{
a data frame  with the censored sample. The attribute
\code{$x} contains the sample values, the attribute \code{$delta}
contains the indicators of non-censoring.

The sample size equals the rows number of p.
}
\description{
\code{gencensg} generates a censored sample from the mixture
of gamma distributions
with mixing probabilities of components given by
the matrix p.
}
\examples{
set.seed(3)
cs <- gencensg(genunifp(50,2),shape=c(2,1),rate=c(0.5,0.5),
shapec=c(1,1),ratec=c(0.3,0.3))
plot(cs$x,col=cs$delta+1)
}
\seealso{
Maiboroda R. Khizanov V.
A modified Kaplan–Meier estimator for a model of mixtures with varying concentrations
Theor. Probability and Math. Statist. 92 (2016), 109-116
}
