% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{genormixt}
\alias{genormixt}
\title{Sample form mixture of normals with warying concentrations.}
\usage{
genormixt(p, mean, sd)
}
\arguments{
\item{p}{matrix (or data frame) of mixing probabilities
with rows corresponding to subjects
and columns coresponding to the mixture components.}

\item{mean}{vector of components' means.}

\item{sd}{vector of components' standard deviations.}
}
\value{
Vector  with the sample.
The sample size equals the rows number of p.
}
\description{
\code{genormixt} generates a sample from the mixture
of normal distributions
with mixing probabilities of components given by
the matrix p.
}
\examples{
  x <- genormixt(genunifp(10,2),c(0,1),c(0.5,0.5))
  plot(x)
}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215
}
