% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{gensampmixt}
\alias{gensampmixt}
\title{Resample form mixture  with warying concentrations.}
\usage{
gensampmixt(p, xs)
}
\arguments{
\item{p}{matrix (or data frame) of mixing probabilities
with rows corresponding to subjects
and columns coresponding to the mixture components.}

\item{xs}{object of class \code{wtsamp} with the weighted
sample. Defines the mixture components' distribution.}
}
\value{
a vector with the resampled sample.
}
\description{
\code{gensampmixt} generates a sample from the mixture
 with mixing probabilities \code{p} and distributions of components given by
the weighted sample \code{xs}.
}
\details{
\code{gensampmixt} uses \code{randwtgen} for sampling
from components with the default value of the option
\code{delta0}.
}
\examples{
set.seed(3)
p <- genunifp(10,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
xs<-cormid(xs) # correct the weights
x_resampled<-gensampmixt(p,xs) # resampled data
}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.

Maiboroda R.  and  Kubaichuk O.
Asymptotic normality of improved weighted empirical distribution functions.
Theor. Probability and Math. Statist. 69 (2004), 95-102
}
