% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{genunifp}
\alias{genunifp}
\title{Generates matrix of mixing probabilities.}
\usage{
genunifp(n, M)
}
\arguments{
\item{n}{number of subjects (rows).}

\item{M}{number of mixing components (columns).}
}
\value{
Matrix p of mixing probabilities with n rows and M columns.
The names of components (columns) are A,B,C...
The names of cases (rows) are 1,2,3,...
}
\description{
\code{genunifp} returns a matrix whose rows are independent
random vectors uniformly distributed over the M-simplex.
}
\examples{
p <- genunifp(10,2)
p

p <- genunifp(1000,3)
plot(p[,3],p[,1],cex=0.2)
}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.
}
