% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescript.R
\docType{data}
\name{hedenf}
\alias{hedenf}
\title{Expression levels of genes from I. Hedenfalk at al. (2001)
study.}
\format{A data frame with 3226 rows corresponding to genes and 22 variables
corresponding to tissue specimens. Names of variables correspond to
the type of tissue:
\describe{
 \item{BRCA1.1}{tissue with the BRCA1 mutation, 1 is the number of
 specimen}
 \item{BRCA1.2}{tissue with the BRCA1 mutation, 2 is the number of
 specimen}
 \item{BRCA1.3}{tissue with the BRCA1 mutation, 3 is the number of
 specimen}
 \item{BRCA1.4}{tissue with the BRCA1 mutation, 4 is the number of
 specimen}
 \item{BRCA1.5}{tissue with the BRCA1 mutation, 5 is the number of
 specimen}
 \item{BRCA1.6}{tissue with the BRCA1 mutation, 6 is the number of
 specimen}
 \item{BRCA1.7}{tissue with the BRCA1 mutation, 7 is the number of
 specimen}
 \item{BRCA2.1}{tissue with the BRCA2 mutation, 1 is the number of
 specimen}
 \item{BRCA2.2}{tissues with the BRCA2 mutation, 2 is the number of
 specimen}
 \item{BRCA2.3}{tissue with the BRCA2 mutation, 3 is the number of
 specimen}
 \item{BRCA2.4}{tissues with the BRCA2 mutation, 4 is the number of
 specimen}
 \item{BRCA2.5}{tissue with the BRCA2 mutation, 5 is the number of
 specimen}
 \item{BRCA2.6}{tissues with the BRCA2 mutation, 6 is the number of
 specimen}
 \item{BRCA2.7}{tissue with the BRCA2 mutation, 7 is the number of
 specimen}
 \item{BRCA2.8}{tissue with the BRCA2 mutation, 8 is the number of
 specimen}
 \item{Sporadic.1}{tissue with sporadic tumors, 1 is the number of
 specimen}
 \item{Sporadic.2}{tissue with sporadic tumors, 2 is the number of
 specimen}
 \item{Sporadic.3}{tissue with sporadic tumors, 3 is the number of
 specimen}
 \item{Sporadic.4}{tissue with sporadic tumors, 4 is the number of
 specimen}
 \item{Sporadic.5}{tissue with sporadic tumors, 5 is the number of
 specimen}
 \item{Sporadic.6}{tissue with sporadic tumors, 6 is the number of
 specimen}
 \item{Sporadic_Meth.BRCA1}{tissue with sporadic tumors in which a
 methilized BRCA1 mutation is present}
}}
\source{
Hedenfalk I. at al. Gene-expression profiles in hereditary breast cancer.
N Engl J Med. 2001 Feb 22;344(8):539-48.
}
\usage{
hedenf
}
\description{
A dataset containing lewel of expression of more than 3226 genes
in breast cancer tissues of 22 patients. The cancer can be caused
by BRCA1 or BRCA2 mutation or can be of Sporadic type.
}
\keyword{datasets}
