% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{IQRw}
\alias{IQRw}
\title{Calculates the interquartile ranges of all components for a weighted sample.}
\usage{
IQRw(xs)
}
\arguments{
\item{xs}{object from class \code{wtsamp} containing the sample
and weights for components' distributions.}
}
\value{
a vector of components' interquartile ranges
}
\description{
\code{IQRw} calculates weighted interquartile ranges of a sample
 using the individual weights of all components.
}
\examples{
set.seed(3)
p <- genunifp(1000,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
IQRw(xs)
}
