% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{KMcdf}
\alias{KMcdf}
\title{Kaplan-Meyer's estimator for CDFs of mixture components}
\usage{
KMcdf(cs)
}
\arguments{
\item{cs}{object of class \code{wtcens} censored sample with
weights representing the distributions
of the censored components.}
}
\value{
object of class \code{wtsamp} containing the sample values
and the cummulative weights for the distributions of the components.
}
\description{
\code{KMcdf} calculates cummulative weigts for the Kaplan-Meyer
estimator of CDFs of mixture components by a censored sample
of class \code{wtcens} and returns an object of class
\code{wtsamp} with the sample values and the calculated weights.
}
\examples{
set.seed(3)
n<-5000
p<-genunifp(n,2)
xs<-gencensg(p,shape=c(1.5,4),rate=c(1,1),shapec=c(1,1),ratec=c(0.3,0.3))
cs<-wtcens(x=xs$x,delta=xs$delta,indiv=lsweight(p))
KMest<-KMcdf(cs)
FKM<-edfgen(KMest,2)
curve(FKM,0,8)
curve(pgamma(x,shape=4,rate=1),col="red",add=TRUE)
}
\seealso{
Maiboroda R. Khizanov V.
A modified Kaplan–Meier estimator for a model of mixtures with varying concentrations
Theor. Probability and Math. Statist. 92 (2016), 109-116
}
