% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{lsweight}
\alias{lsweight}
\title{Calculates minimax weights for components' distributions estimation.}
\usage{
lsweight(p)
}
\arguments{
\item{p}{matrix (or data frame) of mixing probabilities
with rows corresponding to subjects
and columns coresponding to the mixture components.}
}
\value{
matrix (or data frame) of minimax weights of the same structure
as p
}
\description{
\code{lsweight} returns a matrix of individual weights
which correspond to minimax unbiased estimates for CDFs of
mixture components.
}
\examples{
set.seed(3)
p <- genunifp(10,3)
a <- lsweight(p)
t(a)\%*\%p # the result is a unit matrix
}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.
}
