% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{quantilew}
\alias{quantilew}
\title{Calculates the quantiles of all components for a weighted sample.}
\usage{
quantilew(xs, prob)
}
\arguments{
\item{xs}{object from class \code{wtsamp} containing the sample
and weights for components' distributions.}

\item{prob}{the level of the quantiles (one number).}
}
\value{
a vector of components' quantiles.
}
\description{
\code{quantilew} calculates weighted quantiles of a sample
 using the individual weights of all components.
}
\examples{
set.seed(3)
p <- genunifp(1000,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
quantilew(xs,1/3)
}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.
}
