% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{randwtgen}
\alias{randwtgen}
\title{Creates a random number generator according to a weighted sample distribution}
\usage{
randwtgen(xs)
}
\arguments{
\item{xs}{object from class \code{wtsamp} containing the sample
and weights for components' distributions.}
}
\value{
a function \code{f(m,n,delta)}.
The function generates a sample of size \code{n} with the distribution
of the \code{m}-th component of the mixture. \code{delta} is the blurring
parameter.

If \code{delta=0} the generated sample contains values from
\code{xs$x0} sampled with probabilities given by the \code{m}-th
column of \code{xs$indiv}.
If  \code{delta>0} a random variable uniform on [-\code{delta},\code{delta}]
is added to each sampled value.

 The default value of \code{delta} is a half of the minimal distance
 between \code{xs$x0} points.
}
\description{
\code{randwtgen} returns a function wchich produces random samples with the
distribution of a prescribed mixture component. The distribution is estimated
by a weighted sample.
}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.
}
