% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{silvbw}
\alias{silvbw}
\title{Silverman's rule of thumb for bandwidth selection}
\usage{
silvbw(xs, m, delta = 1.7188)
}
\arguments{
\item{xs}{a \code{wtsamp}  object representing
a sample with distributions of different components.}

\item{m}{a number of component for which the density is
estimated.}

\item{delta}{the canonical bandwidth of the kernel
used in the estimate.}
}
\value{
a numeric value of the Silverman's optimal bandwidth.
}
\description{
\code{silvbw} calculates quasi-optimal bandwidth
for kernel density estimate by weighted sample
based on the mixture with varying concentrations approach
}
\details{
The  default value of \code{delta}=1.7188 corresponds to
the Epanechnikov kernel.
}
\examples{
#' @examples
set.seed(3)
p <- genunifp(1000,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
f<-densgen(xs,1) # create the estimator
h<-silvbw(xs,1) # calculates the bandwidth by the Silverman's rule
curve(f(x,h),-3,3) # plot the graph (estimates N(0,1) density)

}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.
}
