% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescript.R
\docType{data}
\name{ukrelect2014}
\alias{ukrelect2014}
\title{Results of Ukrainian Parliament elections 2014}
\format{A data frame with 27 rows corresponding to different
regions of Ukraine and 5 variables:
\describe{
  \item{name}{name of the region}
  \item{code}{code of the region}
  \item{ProEU}{percent of registered voters that voted for
  parties which formed pro-EU coalition  (BPP, Batkivschyna,
   Narodny Front, Radicals and Samopomich)}
   \item{ContraEU}{percent of registered voters that voted for
   the Opposition block}
   \item{Neutral}{percent of registered voters that voted for
   small parties, against all or didn't take part in the voting}
}

(Data on four regions are \code{NA} due to special political situation
at these regions)}
\source{
Official site of Central Election Commission (Ukraine)
\url{http://www.cvk.gov.ua/vnd_2014/}
}
\usage{
ukrelect2014
}
\description{
A dataset containing data on proportions of different
electoral choices in different regions (oblast) in Ukraine
at Parliament (Verhovna Rada) elections in 2014.
}
\keyword{datasets}
