% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{varw}
\alias{varw}
\title{Calculates the variances of all components for a weighted sample.}
\usage{
varw(xs)
}
\arguments{
\item{xs}{object from class \code{wtsamp} containing the sample
and weights for components' distributions.}
}
\value{
a vector of components' variances
}
\description{
\code{varw} calculates weighted variances of a sample
 using the individual weights of all components.
}
\examples{
set.seed(3)
p <- genunifp(1000,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
varw(xs)
}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.
}
