% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{wtcens}
\alias{wtcens}
\title{Constructor for class \code{wtcens}}
\usage{
wtcens(x, delta, cumm = NULL, indiv = NULL)
}
\arguments{
\item{x}{numeric vector containing the sorted sample values.}

\item{delta}{logical vector of non-censoring indicators}

\item{cumm}{matrix (or data frame) of cummulative weights
of components.}

\item{indiv}{matrix (or data frame) of individual weights
of components.}
}
\value{
object of class \code{wtcens}which contains the
following attributes:

\describe{
\item{\code{xo}}{vector of sample values sorted
in the ascending order with \code{-Inf} as the first element.}
\item{\code{deltao}}{vector of non-censoring indicators reordered
at the same order as xo.}
\item{\code{cumm}}{matrix of cummulative weigts reordered
at the same order as xo with 0 at the first row.}
\item{\code{indiv}}{matrix of individual weigts reordered
at the same order as xo.}
}
}
\description{
\code{wtcens} returns an object of S3 class \code{wtcens}
containing the vector of sorted sample values, vector
of indicators of non-censoring
and a set of individual
and/or cummulative weights representing distributions
of different components.
}
