% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start2.R
\name{wtsamp}
\alias{wtsamp}
\title{Constructor for class \code{wtsamp}}
\usage{
wtsamp(x, cumm = NULL, indiv = NULL)
}
\arguments{
\item{x}{numeric vector containing the  sample values.}

\item{cumm}{matrix (or data frame) of cummulative weights
of components.}

\item{indiv}{matrix (or data frame) of individual weights
of components.}
}
\value{
object of class \code{wtsamp} which contains the
following attributes:

\describe{
\item{\code{xo}}{vector of sample values sorted
in the ascending order with \code{-Inf} as the first element.}
\item{\code{cumm}}{matrix of cummulative weigts reordered
at the same order as xo with 0 at the first row.}
\item{\code{indiv}}{matrix of individual weigts reordered
at the same order as xo.}
}

set.seed(3)
p <- genunifp(10,2) # create mixing probabilities
a <- lsweight(p) # calculate minimax weights
# create a weighted sample:
xs <- wtsamp(genormixt(p,c(0,1),c(1,1)),indiv=a)
}
\description{
\code{wtsamp} returns an object of S3 class \code{wtsamp}
containing a sorted sample and a set of individual
and/or cummulative weights representing distributions
of different components.
}
\seealso{
Maiboroda R., Sugakova O. "Statistics of mixtures with varying concentrations 
with application to DNA microarray data analysis". 
Nonparametric statistics (2012) v.24:1, p. 201 - 215.
}
